refactor: move event log writer/reader and file locking to data layer
api reads the Event Log through the new services/run_events facade.
