+++ src/quodeq/data/events/__init__.py
@@ -1,0 +1,1 @@
+"""Event Log adapters (writer/reader) — models stay in core/events/models.py."""
