+++ tests/api/test_run_event_stream.py
@@ -14,7 +14,7 @@ from quodeq.api._run_event_stream import (
     serialize_finding_event,
 )
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 
 
 # ---------------------------------------------------------------------------
