+++ tests/api/test_run_events_routes.py
@@ -10,7 +10,7 @@ from flask import Flask
 
 from quodeq.api._run_events_routes import register_run_events_routes
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 from quodeq.services._evaluations_index import EvaluationsIndex
 from quodeq.services.jobs import JobManager
