+++ tests/data/projection/test_engine_except_breadth.py
@@ -13,7 +13,7 @@ import pytest
 
 from quodeq.data.projection.engine import ProjectionEngine
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 
 
 def _write_one_event(log: Path) -> None:
