+++ tests/data/sqlite/test_findings_repository.py
@@ -6,7 +6,7 @@ from quodeq.core.events.models import (
     JudgmentCreatedEvent,
     JudgmentPayload,
 )
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 from quodeq.data.actions_log import ActionLogWriter
 from quodeq.data.projection.projector import ProjectionResult, Projector
 from quodeq.data.sqlite.connection import open_evaluation_db
