+++ tests/data/test_event_log_reader.py
@@ -6,8 +6,8 @@ from pathlib import Path
 import pytest
 
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.reader import EventLogReader
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.reader import EventLogReader
+from quodeq.data.events.writer import EventLogWriter
 
 
 @pytest.fixture
