+++ tests/integration/test_live_grades_pr1_flow.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 from quodeq.data.sqlite.findings_repository import SqliteFindingsRepository
 from quodeq.services.dismissed import dismiss_finding, restore_finding
