+++ tests/integration/test_sse_run_events_e2e.py
@@ -11,7 +11,7 @@ from flask import Flask
 
 from quodeq.api._run_events_routes import register_run_events_routes
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 
 
 @pytest.fixture
