+++ tests/services/test_deleted.py
@@ -5,7 +5,7 @@ import json
 from pathlib import Path
 
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 from quodeq.data.projection.projector import Projector
 from quodeq.services.deleted import (
     delete_all_dismissed,
