+++ tests/services/test_dismissed_extended.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import pytest
 
 from quodeq.core.events.models import JudgmentCreatedEvent, JudgmentPayload
-from quodeq.core.events.writer import EventLogWriter
+from quodeq.data.events.writer import EventLogWriter
 from quodeq.core.types.finding import Finding, SeverityTally, Totals
 from quodeq.data.projection.projector import Projector
 from quodeq.services.dismissed import (
